TransactionValidator is now solely resposible for updating TipsViewModel
TipsViewModel is responsible for keeping an accurate picture of the solid and non-solid tips. Currently this view is managed within #TransactionValidator.updateStatus() and a separate TipsSolidifier thread (which is not really efficient since it picks one random non-solid tip at a time each 750ms). This leads to a lag in the tips reality. See also issue #761 and #573. With PR #907 in place, we can manage the TipsViewModel fully within TransactionValidator and keep an accurate and up-to-date picture of the current (solid/non-solid) tips.